Add canOpenURL dependency
The `openURL` dependency is useful for opening a URL and checking the
result. However, apps often need to query what URL schemes are supported
on a given device __without actually opening the URL__.

Since the ability to query via the `openURL` dependency is not
supported, a new `canOpenURL` dependency is created. This functionality
is not available on all platforms, so having a separate (but similar)
interface is ideal which also allows the `openURL` dependency to retain
its SwiftUI like semantics.